requirements.txt
Pillow>=7.0.0
matplotlib>=3.0.3
numpy>=1.18.1
opencv-python>=4.2.0.32
pandas>=0.24.2

tensorflow-gpu>= 2.1.0
tf-nightly-gpu>= 2.2.0.dev20200307
tqdm>= 4.43.0
